refactor(All): Fix multiple broken path from refactoring
